+++ commands/stream-mediatext.js
@@ -1,5 +1,12 @@
 const { SlashCommandBuilder } = require("@discordjs/builders");
-const { AttachmentBuilder } = require("discord.js");
+const {
+  AttachmentBuilder,
+  EmbedBuilder,
+  ButtonBuilder,
+  ActionRowBuilder,
+  ButtonStyle,
+  ComponentType,
+} = require("discord.js");
 const path = require("path");
 const fs = require("fs");
 const fetch = require("node-fetch");
@@ -11,6 +18,7 @@ ffmpeg.setFfmpegPath(
 );
 
 const ID_FILE_PATH = path.join(__dirname, "../latest_id.txt");
+const CHANNEL_ID = "1147116498079461466";
 
 let interactionQueue = [];
 let isProcessing = false;
@@ -45,80 +53,179 @@ module.exports = {
     ),
 
   async execute(client, interaction) {
-    await interaction.deferReply();
+    try {
+      await interaction.deferReply();
 
-    interactionQueue.push(interaction);
+      if (!interaction.guild) {
+        await interaction.editReply({
+          content: "This command cannot be used in this context (e.g., DMs).",
+          components: [],
+          ephemeral: true,
+        });
+        return;
+      }
 
-    await interaction.editReply({
-      content:
-        "Your media has been placed in the queue. Please wait while it is being processed.",
-    });
+      interactionQueue.push(interaction);
 
-    if (!isProcessing) {
-      processQueue();
+      await interaction.editReply({
+        content:
+          "Your media has been placed in the queue. Please wait while it is being processed.",
+      });
+
+      processQueue(client);
+    } catch (error) {
+      console.error("Error in execute function:", error);
+      const errorEmbed = new EmbedBuilder()
+        .setColor("#ff0000")
+        .setTitle("Error")
+        .setDescription("An error occurred while processing your request.")
+        .setTimestamp();
+      interaction.followUp({ embeds: [errorEmbed], ephemeral: true });
     }
   },
 };
 
-async function processQueue() {
+async function processQueue(client) {
+  if (isProcessing) return;
+
+  if (interactionQueue.length === 0) return;
+
   isProcessing = true;
 
-  while (interactionQueue.length > 0) {
-    const interaction = interactionQueue.shift();
+  const interaction = interactionQueue.shift();
+  try {
+    const media = interaction.options.getAttachment("media");
+    let duration = interaction.options.getInteger("duration") || 5;
+    const text = interaction.options.getString("text");
 
-    try {
-      const media = interaction.options.getAttachment("media");
-      let duration = interaction.options.getInteger("duration") || 5;
-      const text = interaction.options.getString("text");
+    if (media) {
+      const fileExtension = path.extname(media.url).split("?")[0];
+      const uniqueId = uuidv4();
+      const filePath = path.join(
+        __dirname,
+        "../uploads",
+        `latest_media_${uniqueId}${fileExtension}`
+      );
 
-      if (media) {
-        const fileExtension = path.extname(media.url).split("?")[0];
-        const uniqueId = uuidv4();
-        const filePath = path.join(
-          __dirname,
-          "../uploads",
-          `latest_media_${uniqueId}${fileExtension}`
-        );
+      const response = await fetch(media.url);
+      const buffer = await response.buffer();
+      fs.writeFileSync(filePath, buffer);
 
-        const response = await fetch(media.url);
-        const buffer = await response.buffer();
-        fs.writeFileSync(filePath, buffer);
+      const embed = new EmbedBuilder()
+        .setTitle("Media Validation")
+        .setDescription(
+          `Please approve or reject the media with the following text:\n\n**Text:** ${text}\n**Duration:** ${duration} seconds`
+        )
+        .setColor(0x00ff00)
+        .setImage(`attachment://latest_media_${uniqueId}${fileExtension}`);
 
-        await interaction.editReply({
-          content: `Media received with ID: ${uniqueId} and will be displayed for ${duration} seconds on Streamlabs!`,
-          files: [new AttachmentBuilder(filePath)],
-        });
+      const approveButton = new ButtonBuilder()
+        .setCustomId(`approve_${uniqueId}`)
+        .setLabel("Approve")
+        .setStyle(ButtonStyle.Success);
 
-        if ([".png", ".jpg", ".jpeg"].includes(fileExtension)) {
-          await displayImage(
-            filePath,
-            duration,
-            interaction,
-            fileExtension,
-            uniqueId,
-            text
-          );
-        } else if (fileExtension === ".mp4") {
-          await displayVideo(filePath, interaction, uniqueId, text);
+      const rejectButton = new ButtonBuilder()
+        .setCustomId(`reject_${uniqueId}`)
+        .setLabel("Reject")
+        .setStyle(ButtonStyle.Danger);
+
+      const row = new ActionRowBuilder().addComponents(
+        approveButton,
+        rejectButton
+      );
+
+      const channel = await client.channels.fetch(CHANNEL_ID);
+
+      const message = await channel.send({
+        content: `Media received. Please approve or reject.`,
+        embeds: [embed],
+        components: [row],
+        files: [new AttachmentBuilder(filePath)],
+      });
+
+      const filter = (i) =>
+        i.customId === `approve_${uniqueId}` ||
+        i.customId === `reject_${uniqueId}`;
+      const collector = message.createMessageComponentCollector({
+        componentType: ComponentType.Button,
+        filter,
+        time: 30000,
+      });
+
+      collector.on("collect", async (i) => {
+        if (i.user.id === interaction.user.id) {
+          if (i.customId === `approve_${uniqueId}`) {
+            await i.update({
+              content: "Media approved! It will now be processed.",
+              components: [],
+            });
+
+            if ([".png", ".jpg", ".jpeg"].includes(fileExtension)) {
+              await displayImage(
+                filePath,
+                duration,
+                interaction,
+                fileExtension,
+                uniqueId,
+                text
+              );
+            } else if (fileExtension === ".mp4") {
+              await displayVideo(filePath, interaction, uniqueId, text);
+            }
+          } else if (i.customId === `reject_${uniqueId}`) {
+            await i.update({
+              content: "Media rejected and will not be processed.",
+              components: [],
+            });
+            fs.unlinkSync(filePath);
+          }
+
+          isProcessing = false;
+          processQueue(client);
+        } else {
+          await i.reply({
+            content: "These buttons aren't for you!",
+            ephemeral: true,
+          });
         }
-      } else {
-        await interaction.editReply({
-          content: "Please provide media to send.",
-          ephemeral: true,
-        });
-      }
-    } catch (error) {
-      console.error("Error processing interaction:", error);
-      if (!interaction.replied && !interaction.deferred) {
-        await interaction.editReply({
-          content: "There was an error processing the media. Please try again.",
-          ephemeral: true,
-        });
-      }
-    }
-  }
+      });
 
-  isProcessing = false;
+      collector.on("end", async (collected, reason) => {
+        if (reason === "time") {
+          await message.edit({
+            content: "Validation timed out.",
+            components: [],
+          });
+          if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+          }
+
+          isProcessing = false;
+          processQueue(client);
+        }
+      });
+    } else {
+      await interaction.editReply({
+        content: "Please provide media to send.",
+        ephemeral: true,
+      });
+      isProcessing = false;
+      processQueue(client);
+    }
+  } catch (error) {
+    console.error("Error processing interaction:", error);
+    const errorEmbed = new EmbedBuilder()
+      .setColor("#ff0000")
+      .setTitle("Error")
+      .setDescription("An error occurred while processing the media.")
+      .setTimestamp();
+    if (!interaction.replied && !interaction.deferred) {
+      await interaction.editReply({ embeds: [errorEmbed], ephemeral: true });
+    }
+
+    isProcessing = false;
+    processQueue(client);
+  }
 }
 
 function displayImage(
@@ -130,7 +237,6 @@ function displayImage(
   text
 ) {
   return new Promise((resolve) => {
-
     const data = {
       type: "image-text",
       content: text,
@@ -143,7 +249,9 @@ function displayImage(
     setTimeout(() => {
       try {
         fs.unlinkSync(filePath);
-        fs.unlinkSync(ID_FILE_PATH);
+        if (fs.existsSync(ID_FILE_PATH)) {
+          fs.unlinkSync(ID_FILE_PATH);
+        }
         resolve();
       } catch (error) {
         console.error("Error deleting the image:", error);
@@ -176,7 +284,6 @@ function displayVideo(filePath, interaction, uniqueId, text) {
         try {
           fs.unlinkSync(filePath);
           fs.unlinkSync(ID_FILE_PATH);
-          console.log(`Video deleted after ${videoDuration} seconds.`);
           resolve();
         } catch (error) {
           console.error("Error deleting the video:", error);
